Generalise to a user-selected table
